% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_documentation.R
\name{add_algebraics}
\alias{add_algebraics}
\title{Add algebraic relationships}
\usage{
add_algebraics(form, ...)
}
\arguments{
\item{form}{formula specifying the stochastic differential equation(s) to be added to the system.}

\item{...}{additional formulas similar to \code{form} for specifying multiple equations at once.}
}
\description{
Add algebraic relations.

Algebraic relations is a convenient way to transform parameters in your equations.
In the Ornstein-Uhlenbeck process the rate parameter \code{theta} is always positive, so
estimation in the log-domain is a good idea. Instead of writing \code{exp(theta)} directly
in the system equation one can transform into the log domain using the algebraic relation
\code{add_algebraics(theta ~ exp(logtheta))}. All instances of \code{theta} is replaced
by \code{exp(logtheta)} when compiling the C++ function. Note that you must provide values
for \code{logtheta} now instead of \code{theta} when declaring parameters through
\code{add_parameters}
}
