% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_documentation.R
\name{add_inputs}
\alias{add_inputs}
\title{Declare variables as data inputs}
\usage{
add_inputs(...)
}
\arguments{
\item{...}{variable names that specifies the name of input variables in the defined system.}
}
\description{
Declare variables as data inputs

Declare whether a variable contained in system, observation or observation
variance equations is an input variable. If e.g. the system equation contains
an input variable \code{u} then it is declared using \code{add_inputs(u)}.
The input \code{u} must be contained in the data.frame \code{.data} provided
when calling the \code{estimate} or \code{predict} methods.
}
