% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_documentation.R
\name{add_observation_variances}
\alias{add_observation_variances}
\title{Adding observation variances}
\usage{
add_observation_variances(form, ...)
}
\arguments{
\item{form}{formula class specifying the obsevation equation to be added
to the system.}

\item{...}{additional formulas identical to \code{form} to specify multiple
observation equations at a time.}
}
\description{
Specify the variance of an observation equation.

A defined observation variable \code{y} in e.g. \code{add_observations(y ~ 
h(t,<states>,<inputs>)} is pertubed by Gaussian noise with zero mean and
variance
to-be specified using \code{add_observation_variances(y ~ p(t,<states>,<inputs>)}.
We can for instance declare \code{add_observation_variances(y ~ sigma_x^2}
where \code{sigma_x} is a fixed effect parameter to be declared through
\code{add_parameters}.
}
