% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_documentation.R
\name{add_observations}
\alias{add_observations}
\title{Adding observation equations}
\usage{
add_observations(form, ..., obsnames = NULL)
}
\arguments{
\item{form}{formula class specifying the obsevation equation to be added to the system.}

\item{...}{additional formulas identical to \code{form} to specify multiple observation equations at a time.}

\item{obsnames}{character vector specifying the name of the observation. When the observation left-hand side
consists of more than just a single variable name (when its class is 'call' instead of 'name') it will be
given a name on the form obs__# where # is a number, unless obsnames is provided.}
}
\description{
Define and add a relationship between an observed variable and system states.
The observation equation takes the form

\code{<observation> ~ h(t,<states>,<inputs>) + e)}

where \code{h} is the observation function, and \code{e} is normally
distributed noise with zero mean and variance to be specified. The
observation variable should be present in the data provided when calling
\code{estimate(.data)} for parameter estimation.
}
\examples{
#Specify observation directly as a latent state
add_observations(y ~ x)

Specify observation as the sum of exponentials of two latent states
add_observations(y ~ exp(x1) + exp(x2))
}
