% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_documentation.R
\name{add_parameters}
\alias{add_parameters}
\title{Declare parameters and specify initial values and lower/upper bounds for
optimization procedure.}
\usage{
add_parameters(...)
}
\arguments{
\item{...}{a named vector or matrix as described above.}
}
\description{
Declare which variables that are (fixed effects) parameters in
the specified model, and specify the initial optimizer value, as well as
lower / upper bounds during optimization. There are two ways to declare parameters:
\enumerate{
\item You can declare parameters using formulas i.e. \code{add_parameters( 
theta = c(1,0,10), mu = c(0,-10,10) )}. The first value is the initial
value for the optimizer, the second value is the lower optimization
bound and the third value is the upper optimization bound.
\item You can provide a 3-column matrix where rows corresponds to different
parameters, and the parameter names are provided as rownames of the matrix.
The columns values corresponds to the description in the vector format above.
}
}
