% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_documentation.R
\name{add_systems}
\alias{add_systems}
\title{Adding system state equations}
\usage{
add_systems(form, ...)
}
\arguments{
\item{form}{formula specifying the stochastic differential equation to be
added to the system.}

\item{...}{additional formulas similar to \code{form} for specifying
multiple equations at once.}
}
\description{
Define and add multiple stochastic differential equation governing the
process of individual state variables on the form

\code{d<state> ~ f(t,<states>,<inputs>) * dt + g1(t,<states>,<inputs>) * dw1 
+ g2(t,<states>,<inputs>) * dw2 + ... + gN(t,<states>,<inputs>) * dwN}

where \code{f} is the drift, and \code{g1, g2, ..., gN} are diffusions, with
differential brownian motions dw1, dw2, ..., dwN.
}
\examples{
# Specify Ornstein-Uhlenbeck Process
add_systems(dx ~ theta * (mu - x + u) * dt + sigma * dw)
             
}
