% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.ctsmrTMB.fit}
\alias{plot.ctsmrTMB.fit}
\title{Basic summary of objects of class 'ctsmrTMB'}
\usage{
\method{plot}{ctsmrTMB.fit}(
  fit,
  plot.obs = 1,
  use.ggplot = FALSE,
  extended = FALSE,
  ggtheme = getggplot2theme()
)
}
\arguments{
\item{extended}{logical. if TRUE additional information is printed}
}
\value{
A huge amount of information
}
\description{
Basic summary of objects of class 'ctsmrTMB'
}
