% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.sdemTMB}
\alias{plot.sdemTMB}
\title{Basic summary of objects of class 'sdemTMB'}
\usage{
\method{plot}{sdemTMB}(
  object,
  plot.obs = 1,
  extended = FALSE,
  use.ggplot = FALSE,
  ggtheme = getggplot2theme()
)
}
\arguments{
\item{extended}{logical. if TRUE additional information is printed}
}
\value{
A huge amount of information
}
\description{
Basic summary of objects of class 'sdemTMB'
}
