% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.ctsmrTMB.fit}
\alias{predict.ctsmrTMB.fit}
\title{Predict state estimates on \code{ctsmrTMB.fit} object.}
\usage{
\method{predict}{ctsmrTMB.fit}(
  fit,
  data = NULL,
  x0 = NULL,
  p0 = NULL,
  n.step.ahead = 1,
  ode.timestep = NULL,
  return.state.dispersion = FALSE,
  covariance = FALSE,
  give.only.n.step.ahead = FALSE,
  use.simulation = FALSE,
  sim.timestep = NULL,
  returned.sim.dt = NULL,
  n.sim = 100,
  quantiles = c(0.05, 0.5, 0.95),
  return.full.simulations = FALSE
)
}
\arguments{
\item{fit}{An \code{ctsmrTMB.fit} object from calling \code{estimate} method on \code{ctsmrTMB} object.}

\item{data}{data.frame used to predict and update the model on which \code{fit} is based.}

\item{use.simulation}{Boolean. If TRUE simulates the stochastic model and returns quantiles. if FALSE uses
and returns the solution values for the first and second order moment differential equations. Default: FALSE}

\item{n.sim}{Integer. Number of simulations performed if \code{use.simulation} is TRUE. Default: 100}

\item{quantiles}{The returned state distribution quantiles if \code{use.simulation} is TRUE.}

\item{n.step}{Integer. Number of prediction steps before updating. Default: 1}
}
\value{
Prediction from model
}
\description{
Predict state estimates on \code{ctsmrTMB.fit} object.
}
