% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.sdem.fit}
\alias{print.sdem.fit}
\title{Basic print of objects of class 'sdem'}
\usage{
\method{print}{sdem.fit}(fit)
}
\value{
A huge amount of information
}
\description{
Basic print of objects of class 'sdem'
}
