% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.sdem}
\alias{print.sdem}
\title{Basic print of objects of class 'sdem'}
\usage{
\method{print}{sdem}(object, ...)
}
\value{
A huge amount of information
}
\description{
Basic print of objects of class 'sdem'
}
