% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdemTMB.R
\name{sdemTMB}
\alias{sdemTMB}
\title{R6 Class constructor for sdemTMB}
\value{
A model object of class 'sdemTMB'
}
\description{
Call sdemTMB$new() to construct a new instance of the class object
for specifying a new stochastic state space model.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-sdemTMB-new}{\code{sdemTMB$new()}}
\item \href{#method-sdemTMB-add_systems}{\code{sdemTMB$add_systems()}}
\item \href{#method-sdemTMB-add_observations}{\code{sdemTMB$add_observations()}}
\item \href{#method-sdemTMB-add_observation_variances}{\code{sdemTMB$add_observation_variances()}}
\item \href{#method-sdemTMB-add_inputs}{\code{sdemTMB$add_inputs()}}
\item \href{#method-sdemTMB-add_parameters}{\code{sdemTMB$add_parameters()}}
\item \href{#method-sdemTMB-add_parameters2}{\code{sdemTMB$add_parameters2()}}
\item \href{#method-sdemTMB-add_algebraics}{\code{sdemTMB$add_algebraics()}}
\item \href{#method-sdemTMB-add_constants}{\code{sdemTMB$add_constants()}}
\item \href{#method-sdemTMB-set_initial_state}{\code{sdemTMB$set_initial_state()}}
\item \href{#method-sdemTMB-set_method}{\code{sdemTMB$set_method()}}
\item \href{#method-sdemTMB-set_lamperti}{\code{sdemTMB$set_lamperti()}}
\item \href{#method-sdemTMB-set_loss}{\code{sdemTMB$set_loss()}}
\item \href{#method-sdemTMB-set_modelname}{\code{sdemTMB$set_modelname()}}
\item \href{#method-sdemTMB-set_cppfile_directory}{\code{sdemTMB$set_cppfile_directory()}}
\item \href{#method-sdemTMB-set_compile}{\code{sdemTMB$set_compile()}}
\item \href{#method-sdemTMB-set_silence}{\code{sdemTMB$set_silence()}}
\item \href{#method-sdemTMB-set_map}{\code{sdemTMB$set_map()}}
\item \href{#method-sdemTMB-set_timestep}{\code{sdemTMB$set_timestep()}}
\item \href{#method-sdemTMB-use_hessian}{\code{sdemTMB$use_hessian()}}
\item \href{#method-sdemTMB-build_model}{\code{sdemTMB$build_model()}}
\item \href{#method-sdemTMB-estimate}{\code{sdemTMB$estimate()}}
\item \href{#method-sdemTMB-print}{\code{sdemTMB$print()}}
\item \href{#method-sdemTMB-summary}{\code{sdemTMB$summary()}}
\item \href{#method-sdemTMB-plot}{\code{sdemTMB$plot()}}
\item \href{#method-sdemTMB-clone}{\code{sdemTMB$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-new"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-new}{}}}
\subsection{Method \code{new()}}{
Construct new sdemTMB model object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-add_systems"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-add_systems}{}}}
\subsection{Method \code{add_systems()}}{
Add system equations to model object.

Example:

We add a stochastic differential equation describing the latent random state
variable \code{x} by calling  \code{add_systems(dx ~ theta * (mu - x) * dt 
+ sigma * dw)}. Here \code{theta}, \code{mu} and \code{sigma} are fixed effect
parameters, that must be specified later using the \code{add_parameters}
function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$add_systems(form, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{form}}{formula specifying the stochastic differential equation(s) to be added to the system.}

\item{\code{...}}{additional formulas similar to \code{form} for specifying multiple equations at once.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-add_observations"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-add_observations}{}}}
\subsection{Method \code{add_observations()}}{
Add observation equations to model object.

Example:

Assuming that \code{x} is a state defined through \code{add_systems} with
\code{form = dx ~ ...} then define the relationship between an observation
variable \code{y} and the state \code{x} by calling \code{add_observations(y ~ x)}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$add_observations(form, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{form}}{formula class specifying the obsevation equation to be added to the system.}

\item{\code{...}}{additional formulas identical to \code{form} to specify multiple observation equations at a time.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-add_observation_variances"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-add_observation_variances}{}}}
\subsection{Method \code{add_observation_variances()}}{
Specify observation variance.

Example:

When \code{y} is an observation variable added through \code{add_observation}
then y follows a normal distribution with mean \code{...} and variance which is
specified by calling \code{add_observation_variances(y ~ sigma_y^2)}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$add_observation_variances(form, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{form}}{formula class specifying the obsevation equation to be added to the system.}

\item{\code{...}}{additional formulas identical to \code{form} to specify multiple observation equations at a time.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-add_inputs"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-add_inputs}{}}}
\subsection{Method \code{add_inputs()}}{
Specify system variables that are data inputs.

Example:

If the system equations provided through \code{add_systems(...)} contains
variables say 'Input1', 'Input2' then these must be specified by calling
\code{add_inputs(Input1, Input2)}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$add_inputs(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{additional formulas identical to \code{form} to specify multiple observation equations at a time.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-add_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-add_parameters}{}}}
\subsection{Method \code{add_parameters()}}{
Specify system variables that are fixed effects parameter and
set their (initial) values and lower/upper bounds for the optimization.

Example:
You can provide parameters in two ways:
1.
\code{add_parameters} accepts both a vector of formulas on the form
\code{add_parameters(par1 = c(0,-10,10), par2 = c(5,0,10))} or a matrix of
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$add_parameters(..., parameter.matrix = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{formula class specifying the obsevation equation to be added to the system.}

\item{\code{parameter.matrix}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-add_parameters2"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-add_parameters2}{}}}
\subsection{Method \code{add_parameters2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$add_parameters2(parameters)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-add_algebraics"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-add_algebraics}{}}}
\subsection{Method \code{add_algebraics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$add_algebraics(form, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-add_constants"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-add_constants}{}}}
\subsection{Method \code{add_constants()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$add_constants(form, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-set_initial_state"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-set_initial_state}{}}}
\subsection{Method \code{set_initial_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$set_initial_state(mean, cov)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-set_method"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-set_method}{}}}
\subsection{Method \code{set_method()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$set_method(method)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-set_lamperti"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-set_lamperti}{}}}
\subsection{Method \code{set_lamperti()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$set_lamperti(transform, states = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-set_loss"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-set_loss}{}}}
\subsection{Method \code{set_loss()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$set_loss(loss, c = 3)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-set_modelname"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-set_modelname}{}}}
\subsection{Method \code{set_modelname()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$set_modelname(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-set_cppfile_directory"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-set_cppfile_directory}{}}}
\subsection{Method \code{set_cppfile_directory()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$set_cppfile_directory(directory)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-set_compile"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-set_compile}{}}}
\subsection{Method \code{set_compile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$set_compile(bool)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-set_silence"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-set_silence}{}}}
\subsection{Method \code{set_silence()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$set_silence(bool)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-set_map"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-set_map}{}}}
\subsection{Method \code{set_map()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$set_map(mean, cov)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-set_timestep"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-set_timestep}{}}}
\subsection{Method \code{set_timestep()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$set_timestep(dt)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-use_hessian"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-use_hessian}{}}}
\subsection{Method \code{use_hessian()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$use_hessian(bool)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-build_model"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-build_model}{}}}
\subsection{Method \code{build_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$build_model()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-estimate"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-estimate}{}}}
\subsection{Method \code{estimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$estimate(
  data,
  return.fit = TRUE,
  return.nll = FALSE,
  use.hessian = FALSE,
  ode.timestep = NULL,
  silence = FALSE,
  compile = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-print"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-summary"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-summary}{}}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$summary(correlation = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-plot"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$plot(
  plot.obs = 1,
  use.ggplot = FALSE,
  extended = FALSE,
  ggtheme = getggplot2theme()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdemTMB-clone"></a>}}
\if{latex}{\out{\hypertarget{method-sdemTMB-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdemTMB$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
