% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.ctsmrTMB}
\alias{summary.ctsmrTMB}
\title{Basic summary of objects of class 'ctsmrTMB'}
\usage{
\method{summary}{ctsmrTMB}(object, correlation = FALSE)
}
\value{
A huge amount of information
}
\description{
Basic summary of objects of class 'ctsmrTMB'
}
