% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.sdem}
\alias{summary.sdem}
\title{Basic summary of objects of class 'sdem'}
\usage{
\method{summary}{sdem}(object, correlation = FALSE)
}
\value{
A huge amount of information
}
\description{
Basic summary of objects of class 'sdem'
}
