% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.sdemTMB}
\alias{summary.sdemTMB}
\title{Basic summary of objects of class 'sdemTMB'}
\usage{
\method{summary}{sdemTMB}(object, correlation = FALSE)
}
\value{
A huge amount of information
}
\description{
Basic summary of objects of class 'sdemTMB'
}
